requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
colorlog==6.9.0
dnspython==2.7.0
fastapi==0.115.8
fastapi-limiter==0.1.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
motor==3.7.0
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
pymongo==4.11.1
python-telegram-bot==21.10
redis==5.3.0b5
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
